deliveries/alisboa/Desafio 10 - Final/backend/requirements.txt
sqlalchemy
jinja2
connexion
pyyaml